Add support for AIX systems
As it turns out, all *nix systems are not created equal. As such, while
AIX does have a group with GID of 0, that group is not known as 'root'.
Instead, AIX lovingly knows this group as the 'system' group. This means
the resolv_conf module could not manage AIX systems as it is attemping
to set ownership of the /etc/resolv.conf file to 'root:root'.

This commit changes the enforced ownership to use '0:0' instead of
'root:root'. The spec tests are also updated to reflect this change.